+++ tiger-proxy/src/main/java/de/gematik/test/tiger/exception/TigerProxyModificationException.java
@@ -1,0 +1,7 @@
+package de.gematik.test.tiger.exception;
+
+public class TigerProxyModificationException extends RuntimeException {
+    public TigerProxyModificationException(String s) {
+        super(s);
+    }
+}