+++ tiger-testenv-mgr/src/main/java/de/gematik/test/tiger/testenvmgr/DockerMgr.java
@@ -22,6 +22,7 @@ import com.github.dockerjava.api.exception.DockerException;
 import com.github.dockerjava.api.exception.NotModifiedException;
 import com.github.dockerjava.api.model.ContainerConfig;
 import com.github.dockerjava.api.model.PullResponseItem;
+import com.github.dockerjava.api.model.ResponseItem;
 import de.gematik.test.tiger.common.OsEnvironment;
 import de.gematik.test.tiger.testenvmgr.config.CfgEnvSets;
 import de.gematik.test.tiger.testenvmgr.config.CfgServer;
@@ -56,10 +57,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.time.Duration;
 import java.time.temporal.ChronoUnit;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
+import java.util.*;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
 
@@ -81,7 +79,7 @@ public class DockerMgr {
 
         pullImage(imageName);
 
-        final GenericContainer<?> container = new GenericContainer<>(testContainersImageName);
+        final var container = new GenericContainer<>(testContainersImageName); //NOSONAR
         try {
             InspectImageResponse iiResponse = container.getDockerClient().inspectImageCmd(imageName).exec();
             final ContainerConfig containerConfig = iiResponse.getConfig();
@@ -139,13 +137,12 @@ public class DockerMgr {
                 .withName("tiger." + server.getHostname()).exec();
             containers.put(server.getHostname(), container);
             final Map<Integer, Integer> ports = new HashMap<>();
-            // TODO for now we assume ports are bound only to one other port on the docker container
+            // TODO TGR-282 LO PRIO for now we assume ports are bound only to one other port on the docker container
+            // maybe just make clear we support only single exported port
             container.getContainerInfo().getNetworkSettings().getPorts().getBindings().entrySet().stream()
                 .filter(entry -> entry.getValue() != null)
                 .forEach(entry -> ports.put(entry.getKey().getPort(), Integer.valueOf(entry.getValue()[0].getHostPortSpec())));
             server.getDockerOptions().setPorts(ports);
-
-
         } catch (final DockerException de) {
             throw new TigerTestEnvException("Failed to start container for server " + server.getHostname(), de);
         }
@@ -231,7 +228,13 @@ public class DockerMgr {
             .exec(new ResultCallback.Adapter<PullResponseItem>() {
                 @Override
                 public void onNext(PullResponseItem item) {
-                    log.debug(item.getStatus() + " " + (item.getProgressDetail() != null ? item.getProgressDetail().getCurrent() : ""));
+                    if (log.isDebugEnabled()) {
+                        log.debug("{} {}", item.getStatus(),
+                            Optional.ofNullable(item.getProgressDetail())
+                                .map(ResponseItem.ProgressDetail::getCurrent)
+                                .map(Object::toString)
+                                .orElse(""));
+                    }
                 }
 
                 @Override
@@ -298,7 +301,7 @@ public class DockerMgr {
                 content += "grep -q \"host.docker.internal\" /etc/hosts || "
                     + "echo \" \" >> /etc/hosts && echo \"" + hostip + "    host.docker.internal\" >> /etc/hosts\n";
 
-                // TODO reactivate once we have docker 20 on maven nodes
+                // TODO TGR-283 reactivate once we have docker v20 on maven nodes
                 //  container.withExtraHost("host.docker.internal", "host-gateway");
                 content += "echo HOSTS:\ncat /etc/hosts\n";
             } else {