+++ tiger-testenv-mgr/src/test/java/de/gematik/test/tiger/testenvmgr/config/TestTigerTestEnvMgr.java
@@ -48,7 +48,7 @@ public class TestTigerTestEnvMgr {
 
     @BeforeAll
     public static void proxySettings() {
-        // TODO check whether to remove once the Jenkinsfile has been merged to master
+        // TODO TGR-286 check whether to remove once the Jenkinsfile has been merged to master
         if (System.getenv("PROXY_HOST") != null) {
             log.info("Applying Jenkins proxy env vars! " +
                 System.getenv("PROXY_HOST") + ":" + System.getenv("PROXY_PORT"));
@@ -412,6 +412,6 @@ public class TestTigerTestEnvMgr {
         Thread.sleep(2000);
     }
 
-    // TODO check pkis set, routings set,....
+    // TODO TGR-286 check pkis set, routings set,....
 
 }